refactor: remove local build script and update imports to reference source files directly
