+++ docs/.gitignore
@@ -1,5 +1,6 @@
 # build output
 dist/
+src/lib/capo.js
 # generated types
 .astro/
