+++ docs/src/components/CapoClient.astro
@@ -5,7 +5,7 @@ import VirtualConsole from "./VirtualConsole.astro";
 ---
 
 <script>
-  import { run } from "../lib/capo.js";
+  import { run } from "../../../src/web/capo.js";
   import { fetchGist, exchangeCodeForToken, createGist } from "../lib/gist.js";
   import { highlightHTML } from "../lib/CapoWidgetElement.js";
   import type { VirtualConsole } from "../lib/VirtualConsole.js";
